# DOC: optimize: make `lsq_linear` example smaller