Before the change, CircleCI periodically times out (in https://github.com/scipy/scipy/pull/21437), and

```
$ python smoke-docs -s optimize --durations 10
...
10.23s call     build-install/lib/python3.12/site-packages/scipy/optimize/__init__.py::scipy.optimize.lsq_linear
4.46s call     build-install/lib/python3.12/site-packages/scipy/optimize/__init__.py::scipy.optimize.differential_evolution
1.35s call     build-install/lib/python3.12/site-packages/scipy/optimize/__init__.py::scipy.optimize.least_squares
0.87s call     build-install/lib/python3.12/site-packages/scipy/optimize/__init__.py::scipy.optimize.dual_annealing
```

<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#writing-the-commit-message.
However, please only include an issue number in the description, not the title,
and please ensure that any code names containing underscores are enclosed in backticks.

Depending on your changes, you can skip CI operations and save time and energy: 
http://scipy.github.io/devdocs/dev/contributor/continuous_integration.html#skipping

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->

#### Reference issue
<!--Example: Closes gh-WXYZ.-->

#### What does this implement/fix?
<!--Please explain your changes.-->

#### Additional information
<!--Any additional information you think is important.-->