@ev-br 2024-08-24
One other outlier is `stats.multiscale_graphcorr`:

```
$ python dev.py smoke-docs -s stats --durations 10

....

===================================================================== slowest 10 durations =====================================================================
23.01s call     build-install/lib/python3.12/site-packages/scipy/stats/__init__.py::scipy.stats.multiscale_graphcorr
3.04s call     build-install/lib/python3.12/site-packages/scipy/stats/__init__.py::scipy.stats.recipinvgauss
2.28s call     build-install/lib/python3.12/site-packages/scipy/stats/__init__.py::scipy.stats.studentized_range
2.06s call     build-install/lib/python3.12/site-packages/scipy/stats/__init__.py::scipy.stats.gausshyper
1.86s call     build-install/lib/python3.12/site-packages/scipy/stats/__init__.py::scipy.stats.rel_breitwigner
1.84s call     build-install/lib/python3.12/site-packages/scipy/stats/__init__.py::scipy.stats.foldcauchy
1.80s call     build-install/lib/python3.12/site-packages/scipy/stats/sampling.py::scipy.stats.sampling.NumericalInverseHermite
1.70s call     build-install/lib/python3.12/site-packages/scipy/stats/sampling.py::scipy.stats.sampling.NumericalInversePolynomial
1.47s call     build-install/lib/python3.12/site-packages/scipy/stats/__init__.py::scipy.stats.kappa4
1.40s call     build-install/lib/python3.12/site-packages/scipy/stats/__init__.py::scipy.stats.goodness_of_fit
```