@ev-br 2024-08-24
Not really. The stats one deserves a look from a subject matter expert IMO, I'm not comfortable assessing if the sample sizes can just change or a more invasive change is needed to keep the example sensible.