@ev-br 2024-08-24
Uhm, sorry, to clarify: I think the `multiscale_graphcorr` could/should be looked at by a subject matter expert, so there's no need to lump it together with `lsq_linear` of this PR. So this PR is ready to go in, IMO.