@WarrenWeckesser 2024-08-24
Thanks, the change makes sense.

The example output should be updated to reflect the change in size.  It looks like this causes the typical number of iterations to be lower, and the typical costs differ by an order of magnitude compared to the larger case.  The numbers will vary, but they should probably continue to reflect a "typical" result from the input sizes used in the example.