@WarrenWeckesser 2024-08-24
I don't recall the scope of `# may vary`.  It looks like the occurrence of `# may vary` in an output section applies to all lines in that output, and not just the one line where is occurs.  Is that correct?