@ev-br 2024-08-24
Yes, `# may vary` should apply to the whole output, multiline or not. Dropping a breakpoint at https://github.com/scipy/scipy_doctest/blob/main/scipy_doctest/impl.py#L285 gives:


```
(Pdb) p want
'# may vary\nThe relative change of the cost function is less than `tol`.\nNumber of iterations 16, initial cost 1.5039e+04, final cost 1.1112e+04,\nfirst-order optimality 4.66e-08.\n'
(Pdb) p got
'The relative change of the cost function is less than `tol`.\nNumber of iterations 16, initial cost 1.4910e+04, final cost 1.1015e+04, first-order optimality 6.95e-08.\n'
```

(this is on main not this PR for a random reason)